[apex] Restore binary compat for ApexCpdLexer
The #tokenize() method has been moved to the super class. This
is not a problem. But the super class defines the method as
final, which means it can't be overridden.

For now, we simply delegate to the new version of the apex
lexer. With PMD 8, we can merge the two implementations
again.
